AI编程社团/AI资讯项目/requirements.txt
feedparser>=6.0.10
pyyaml>=6.0
requests>=2.31.0
schedule>=1.2.0
